Change encoding for production database
This parameter was initially set to 'utf8mb4'. Due to the default
configuration that was generated when the project was crated using
'-d=mysql' flag to set mysql as the default database. It was changed to
'utf8' instead, since production environment is running a postgresql
database and 'utf8mb4' is not a valid unicode for it.